+++ libs/mat4lib/srcs/vect_scale.c
@@ -6,13 +6,13 @@
 /*   By: gmachena <gmachena@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/19 15:55:21 by gmachena          #+#    #+#             */
-/*   Updated: 2020/02/19 15:56:14 by gmachena         ###   ########.fr       */
+/*   Updated: 2020/03/05 10:16:12 by gmachena         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "mat4lib.h"
 
-t_vect3	vect3_scale(t_vect3 v, double scale)
+t_vect3	vect3_scale(t_vect3 v, float scale)
 {
     v.v[0] *= scale;
     v.v[1] *= scale;
@@ -20,7 +20,7 @@ t_vect3	vect3_scale(t_vect3 v, double scale)
     return (v);
 }
 
-t_vect4	vect4_scale(t_vect4 v, double scale)
+t_vect4	vect4_scale(t_vect4 v, float scale)
 {
     v.v[0] *= scale;
     v.v[1] *= scale;
